Refactor the identify method
