+++ tests/identify.py
@@ -1,0 +1,5 @@
+from app.main import identify
+
+if __name__ == "__main__":
+    upload_file()
+    identify()
